quality: solve all gcc compiler warnings (#51)
* quality: solve all gcc compiler warnings

- Solve all implicit numeric conversion warnings
- Solve all useless casting warnings
- Solve all implicit this capture warnings which deprecated in c++20
- Solve all unused variables warnings
- Solve multiline comment warning
- Solve qt deprecated warnings
- Solve data members intialization order warning
- Solve function parameter not used warning

Ticket: SOUR-76

* Docker: fix missing xcb platform plugin coredump

Ticket: SOUR-76

+++ src/lib/component/controller/UndoRedoController.cpp
@@ -539,13 +539,13 @@ void UndoRedoController::updateHistory() {
       }
 
       if(historyListMatches.size() == historyListSize && currentIndex != -1 &&
-         currentIndex - m_historyOffset != historyListSize - 1) {
+         static_cast<size_t>(currentIndex) - m_historyOffset != historyListSize - 1) {
         break;
       }
     }
   }
 
-  getView()->updateHistory(historyListMatches, currentIndex - m_historyOffset);
+  getView()->updateHistory(historyListMatches, static_cast<size_t>(currentIndex) - m_historyOffset);
 }
 
 void UndoRedoController::dump() const {
